tensorflow support update
